Dedup: extract shared Cloudflare capture pipeline (#6297)
